astakos: Make limit_on_members_number a bigint
We need bigint in order to represent infinite as 2**63-1.

+++ snf-astakos-app/astakos/im/models.py
@@ -1313,7 +1313,7 @@ class ProjectApplication(models.Model):
     end_date = models.DateTimeField()
     member_join_policy = models.IntegerField()
     member_leave_policy = models.IntegerField()
-    limit_on_members_number = models.PositiveIntegerField(null=True)
+    limit_on_members_number = models.BigIntegerField(null=True)
     resource_grants = models.ManyToManyField(
         Resource,
         null=True,
